Renamed ScoreObjectSelector to ScoreElementSelector.
(The word 'object' is a really terrible one to have embedded in any custom class or function,
given that the word is reserved in Python and shouldn't be used as a loop argument, for example.)

Used the new --classes flag of abj-dev rename to make the change. Everything appears to work.

The new output like this ...

    Replaced 1042 instances over 497 lines in 59 files.

... is both nicer and more accurate than the previous output.

I'm also surprised at the number of replacement instances found. Maybe because the new
version of the script is catching all the cases in the reST files? And maybe also
because I was never aware of the total number of replacement instances in previous usage
of the script (because previous versions of the script reported this number incorrectly!).