Removed 'count' property from ScoreElementSelection.
The property was only a stub.

The property was aiming to afford functionality similar to these three functions in the public API:

    componenttools.iterate_namesakes_forward_from_component
    componenttools.iterate_thread_forward_from_component
    componenttools.iterate_timeline_forward_from_component

The idea had two parts. First pick out a score element with a selector. Then specify
the number of similar elements to pick out using the first score element as a starting point.

This is still a very good idea. But it will be modeled in a different way.