Renamed specificationtools.ScoreElementSelector to specificationtools.ComponentSelector.
The new name is better because there is now a distinction in codebase between
components (which are the usual notes, rests, chords, tuplets, voices, staves of
all the previous years of Abjad development) ands 'background elements' (which
are the new segments and divisions particular to the specificationtools package,
together with any measures that may be used in a 'backgrounded' way).

Because the codebase now features a distinction between components and background
elements, it makes sense to implement some selectors for components and some
other selectors for background elements.