Renamed TimespanInequalityClass to TimespanInequalityTaxon.
The word 'class' is reserved in Python.

What is the relationship between a TimespanInequalityTaxon and an actual TimespanInequality?

A timespan inequality taxon is something like "expr.start == timespan.start".
This means something like "all expressions that start when some as-yet-unspecified 
reference timespan starts".

We add an actual timespan to change this taxon into an actual inequality.
So "expr.start == timespan.start" can combine with specificationtools.SegmentSelector('red').timespan.
This gives "all expression that start when the timespan of segment 'red' starts".