Renamed one selector class.
    selectortools.ComponentSelector ==>
    selectortools.CounttimeComponentSelector

What does 'counttime' mean?

This commit introduces 'counttime' as a possible (nonoppositional) label
to mean 'noncontext'. Counttime components are leaves, tuples, measures
and naked containers. What do these components have in common?

    * They are not contexts
    * They exist at levels of the score tree below the voice
    * They can nest with each other in interesting ways
    * They are essentially the componens that set up the rhythmic tree
      that realworld performers will be obliged to count during performance
      (ie, these components are mostly notes, rests, tuplets and measures)

Naming this concept is hard. But it turns out to be important. Because several
of the selectortools classes will pick out counttime components but will not
be used to pick out context component. And it would be confusing to have
selector classes named things like ComponentSelector when such a class would
refuse to operate on context components and would instead operate only
on counttime components. So better to have a CounttimeComponentSelector class
that operates on counttime components.