Add four new selector classes:
    * selectortools.CounttimeContainerSelector
    * selectortools.CounttimeContainerSliceSelector
    * selectortools.LeafSelector
    * selectortools.LeafSliceSelector

The current implementations are stubs.

The classes inherit from CounttimeComponentSelector and CounttimeComponentSliceSelector.

Subclassing this way makes the code easier to read.

Subclassing this way also better affords the right type of nesting in the system.
Why? Because it is uniquely the new CounttimeContainerSelector that will be able
to serve as the reference for all four classes recursively. (Because it only
makes sense to slice from a container and not from a leaf.)

Examples to be added to all four new classes. And the base classes will become abstract.