Search-and-replace fix to previous commit.
There's a bug somewhere in "abj-dev rename -C".

In this case the bug surfaced when renaming ...
    
    selectortools.MeasureSliceSelector ==>
    selectortools.BackgroundMeasureSliceSelector

... with ...
    
    selectortools.BackgroundBackgroundMeasureSliceSelector

... resulting from the rename.

Hanve't hunted the bug down. But when it shows up, it always shows
up in this way of reduplicating the first camelcase word of a class name.

(Bug occurred early this week but held off for a few more class
renames to see if it would reappear.)