Working on count-ratio item-selector integration.
What does this mean?

This means the ability to make a setting for a *fractional part* of a score segment.

The test case is this:

    ### RATIO OF BACKGROUND MEASURES TEST ###

    template = scoretemplatetools.GroupedRhythmicStavesScoreTemplate(n=1)
    score_specification = specificationtools.ScoreSpecification(template)

    segment = score_specification.append_segment()
    segment.set_time_signatures(segment, [(4, 8), (3, 8)])

    selector = segment.select_ratio_of_background_measures((1, 1), 0)
    segment.set_divisions(selector, [(3, 16)], truncate=True)

    selector = segment.select_ratio_of_background_measures((1, 1), 1)
    segment.set_divisions(selector, [(5, 16)], truncate=True)

    segment.set_rhythm(segment, library.thirty_seconds)

    score = score_specification.interpret()

    ### END TEST ###

The test case divides the (trivial) two measures of the example into a 1:1
ratio and then sets (3, 16) divisions for the first part and (5, 16) divisions
for the second part. The 1:1 partitioning here is trivial: the first part
equals measure 1 and the second part equals measure 2.

The test doesn't work as of this commit. But it's extremely close.

Changes in this commit:

Replaced namedtuples with full-fledged classes:

    * interpretertools.SegmentDivisionToken
    * interpretertools.RegionDivisionToken

Remove unused 'timespan' keyword from:
    
    SegmentSpecification.get_divisions_value_with_fresh_and_truncate()
    SegmentSpecification.get_rhythm_value()
    ContextDictionary.get_settings()
    ContextProxy.get_setting()
    ContextProxy.get_settings()

Changed the way that resolved settings context dictionaries make use of context proxies.
Previously context proxies stored zero or one resolved setting for a given attribute.
Now context proxies store a list of zero or more resolved settings for a given attribute.