Renaming all interpretertools -Token classes to -Command classes.
The word 'token' is used in a couple of vaguely conflicting ways in the codebase.

In the mainline, the timetokentools package implements one idea of what a token is.

But in the experimental branch, the idea of a token in interpretertools was conflicting.

This commit begins the process of changing interpretertools tokens to commands.