Renamed one specificationtools class:
    specificationtools.RegionDivisionList ==>
    specificationtools.DivisionRegionDivisionList

The rename reflects a new understanding.

The new understanding is that the process of interpretation creates
several different types of 'region'.

Interpretation creates division regions. Interpretation creates rhythm regions.
Interpreation will eventually create pitch-class regions and pitch regions
and articulation regions.

A region is just a special type of timespan.

A region is a timespan to which a single command applies.