Added new sequencetools function:
    * sequencetools.pair_duration_sequence_elements_with_input_pair_values()

Example:

    >>> duration_sequence = [10, 10, 10, 10]
    >>> input_pairs = [('red', 1), ('orange', 18), ('yellow', 200)]

    >>> sequencetools.pair_duration_sequence_elements_with_input_pair_values(
    ... duration_sequence, input_pairs)
    [(10, 'red'), (10, 'orange'), (10, 'yellow'), (10, 'yellow')]

The basic idea behind the function is model which input pair value is in effect 
at the start of each element in duration_sequence.

The example above accepts a duration sequence equal to [10, 10, 10, 10].
This models four objects each with a duration of 10.
This means that the start offsets of these four objects are [0, 10, 20, 30].

The example above accepts input pairs equal to [('red', 1), ('orange', 18), ('yellow', 200)].
This means that the value 'red' is in effect for a duration of 1,
that the value 'orange' is in effect for a duration of 18,
and that the value 'yellow' is in effect for a duration of 200.

The output of the example above is [(10, 'red'), (10, 'orange'), (10, 'yellow'), (10, 'yellow')].
This means that the first element starts while 'red' is active,
that the second element starts with 'orange' is active,
and that the last two elements start while 'yellow' is active.