Added new division list class:
    * interpretationtools.RhythmRegionDivisionList

What is a rhythm region?
A rhythm region is an uninterrupted block of time to which a rhythm-maker applies.

What is a rhythm region division list?
A rhythm region division list is a special type of division list.
A rhythm region division list is the division list passed as
input to a rhythm-maker.