Migrated all unpacking logic into ConcreteInterpreter.
What does unpacking logic do?

Unpacking means to change a MultipleContextSetting into zero or more SingleContextSettings.

So unpacking is a bit like 'exploding' multiple-context settings.

If this behavior were specification-time behavior then this logic would belong
in one of the specification classes.

But because this behavior is interpretation-time behavior this logic belongs
in one of the interpreter classes.