Made SegmentSpecification.set_attribute() method private.
The new name is SegmentSpecification._set_attribute().

The SegmentSpecification._set_attribute() method is a generalized
method to create MultipleContextSetting objects. This is powerful.
But can also be annoying for composers to use. So the design idea
is that ScoreSpecification and SegmentSpecification will implement
an extensive inventory of setter methods that will do everything
the composer would like. This will make score-building files
largely self-documenting. This also requires that developers
implement, maintain and extend the inventory of setter methods
over time.