Added ConcreteInterpreter.single_context_timespan_selector_to_duration() method.
The purpose of the method is to get the duration of a single-context timespan selector.

The method is not yet generalized.
The current implementation of the method handles only two cases.
The first case is that in which the selector ultimately references a segment.
The second case is that in which the selector ultimately references a slice of background measures.

It is the second (background measure slice) case that starts to extend interpretation
to be able to handle several differernt division regions in a single segment.