Removed SingleContextSetting initializer polymorphism. Removed MultipleContextSetting initializer polymorphism.
The two classes previously allowed initialization from other instances
of the same class. This commit removes that affordance.

This type of polymorphism has proved both useful and clean when
implemented on objects in the Abjad mainline.

But the extra code implemented on the Setting classes here in the
experimental line simply cluttered the class implementations without
providing any truly better functionality.

It is also now the case that the initializer signatures of
both classes are now clearly visible in the docs.