Extended UninterpretedDivisionCommand with stub 'context_name' property.
Uninterpreted division comands need to carry around information about
the context from which they derive. The reason for this is that
two or more division commands may overlap. (That means that two or
more division commands may be specified for part of the same timespan.)
The interpreter must know how to break the tie when two or more
division comamnds overlap. Knowing how to break such a tie
requires access to the source contexts from which division commands.
(And the reason for this is that commands arising from lower-level
contexts will override commands arising from higher-level contexts.
The intuition here is that a division command specified globally for
the entire score is less important than a division command specified
more surgically for just a single voice.)