Extended interpreter commands.
There are currently three concrete interpreter commands:

    * interpretertools.UninterpretedDivisionCommand
    * interpretertools.RegionDivisionCommand
    * interpretertools.RhythmCommand

Interpreter commands basically say "apply this division-maker to this region"
or "apply this rhythm-maker to this region". So an interpreter command
is a bit like the object representation of a call to a division-making function
or like the object representation of a call to a rhythm-making function.

As of this commit all interpreter commands now implement a read-only public
'start_segment_name' property. The purpose of the property is to track
the segment which gave rise to the interpreter command. This is useful
for sorting commands by segment.