Extended objects that initialize with timespans.
Objects that except timespans at initialization are now smart enough
to look for a public 'timespan' property and use that property
when available.

This confers a benefit regarding selectors. All selectors implment
a public timespan property. This means that you can pass selectors
around the system and have them be treated as timespans transparently.