Regularizing selector interface.
The word 'index' has too many meanings in the system.
The word should be reserved for the iterable.index() method
and for the iterable[n] argument.

The interpreter-related packages introduce two new uses
of 'index'. The first of these is in relation to ratio selectors
(ie, which item or part of a ratio does the selector pick out?).
The second of these is in relation to segments (ie, as a generalized
term that covers both segment name and segment index).

The strategy will be to change ratio selectors to work with an 'item'
and to change segment selectors to work with an 'identifier'.