Renamed selectortools.selector_to_segment_index(). The new name is selectortools.selector_to_segment_identifier().
Explanation:

A score specification comprises a list of zero or more segment specifications.
This means that every segment specification has a nonnegative integer index
(just like a list element). But not that segment specifications can also
be given a (string) name. So segment specifications can be identified either
by (string) name or by (nonnegative integer) index.

The term 'identifier' is introduced here to cover both of these ways
of potentially identifying a segment.