Optimized more segment identifier-determination logic.
The primary design goal is that all selectors be able to
determine their segment identifier recursively.

A secondary design goal is that each selector delegate
the determination of segment identifer to only
one aggregated component. That is, the secondary design
goal is to limit dot-chained references to at most
three objects chained in a row (and let recursion take
care of the rest of it).