Improved ConcreteInterpreter.store_single_context_settings(new_settings, clear_persistent_first=False).
The method loops through every new setting, resolves the seting
and then stores the setting.

Previously the method consulted the clear_persistent_first flag on every setting.
Now the method consults the clear_persistent_first flag on only the first setting.

The reason for the change is that the method now treats a batch of settings as a batch.
That is good because existing settings sitting up in the global registry of
settings should be cleared out at the beginning of a transaction but not *during*
a transaction.