Renamed four measuretools functions:
    measuretools.fill_measures_in_expr_with_meter_denominator_notes()
    measuretools.move_prolation_of_full_measure_tuplet_to_meter_of_measure()
    measuretools.multiply_contents_of_measures_in_expr_and_scale_meter_denominators()
    measuretools.scale_measure_by_multiplier_and_adjust_meter()

    ==>

    measuretools.fill_measures_in_expr_with_time_signature_denominator_notes()
    measuretools.move_full_measure_tuplet_prolation_to_measure_time_signature()
    measuretools.multiply_contents_of_measures_in_expr_and_scale_time_signature_denominators()
    measuretools.scale_measure_and_adjust_time_signature()

The use of 'meter' is deprecated everywhere in the codebase.

The reason for this is that Abjad has no formal meter model.

Abjad does have a formal model of time signatures and also measures.

So use 'time_signature' everywhere in preference to 'meter'.

(At least until such a time as we create a formal meter model
that differs from that of time signature.)

NOTE: this change affects end-user syntax and must be noted in the 2.10 release notes.