Beautified tuplet example in RhythmTree parsing overview:
Explanation: there is a 'docs=True' keyword available for use with show().
The keyword is set in the 'why LilyPond is right for Abjad' page in the docs.
The keyword doesn't do all that much. But it does apply some LilyPond styling.
And one element of LilyPond styling is some tuplet bracket padding.
So making this setting fixed the tuplet bracket collisions in the example.

Also put the example in a rhythmic staff. This is always nice when the example
is an unpitched rhythmic figure. (Especially one with tuplet brackets.)