Renamed requesttools.resolve_request_offset_and_count().
The new name is requesttools.apply_request_transforms().

Explanation: 'offset' and 'count' are two different request transforms.
But now a new 'reverse' transform is being implemented. And conceivably
more transforms will be implemented in the coming months. So
this commit generalizes the name of the helper function responsible
for the application of these transforms.