Removed four deprecated tuplettools functions:
    tuplettools.make_augmented_tuplet_from_duration_and_proportions_and_avoid_dots
    tuplettools.make_augmented_tuplet_from_duration_and_proportions_and_encourage_dots
    tuplettools.make_diminished_tuplet_from_duration_and_proportions_and_avoid_dots
    tuplettools.make_diminished_tuplet_from_duration_and_proportions_and_encourage_dots

Use tuplettools.make_tuplet_from_duration_and_proportions() instead.

NOTE: this change affects end-user syntax and must be noted in the 2.10 release notes.

Note, too, that this commit completes the deprecation of nonstandard uses of 'and' in the codebase.

All remaining uses of 'and' in the names of functions in the API are now acceptable
and no longer suggest replacement with boolean keywords.