Added stub timespantools.TimespanInventory class.
The class will eventually function as a smart collection of timespans.

The class will include methods to retrieve timespans starting before,
with or after or a certain offset, and so on.

Any object implementing the 'timespan interface' can be stored
in a timespan inventory.

The 'timespan interface' is satisfied when an object implements
'start_offset' and 'stop_offset' properties.

Timespan inventory subclasses will be possible. A DivisionRegionDivisionListInventory
is particuarly likely because the class can then implement special retrieval-plus-slicing
methods peculiar to (and useful for working with) division list objects.