Extended TimespanInequality with optional 'expr' property.
Example. Make timespan inequality indicating that offset 7/16 happens during segment 'red':

        >>> selector = selectortools.SingleSegmentSelector(identifier='red')
        >>> timespan_inequality = timespantools.expr_happens_during_timespan(
        ...     selector, expr=durationtools.Offset(7, 16))

        >>> z(timespan_inequality)
        timespantools.TimespanInequality(
            timespantools.TimespanInequalityTemplate('t.start <= expr.start <= expr.stop <= t.stop'),
            timespantools.SingleSourceTimespan(
                selector=selectortools.SingleSegmentSelector(
                    identifier='red'
                    )
                ),
            expr=durationtools.Offset(7, 16)
            )

This is leading up to implementing __call__ on TimespanInequality objects.

TimespanInequality objects that are 'fully loaded' with both an evaluable timespan
and an evaluable expr will then be callable. The result will be a boolean indcating
whether expr compares to the reference timespan in the way initialized in the object.