Added stub labeltools package.
The 28 label functions currently housed in the other packages 
(primarily leaftools and verticalitytools) will migrate to labeltools.

Background: finishing the work of recently rewiring all the import
statements gives us one of the best views into package architecture.
In particular, all of the packages implemented recently (in the last
year or even two) exhibit a relatively clean separation from the core
packages in the system. (Dependencies between core packages of the
system are harder to work out but this is to be at least somewhat expected.)

Extracting the couple dozen label functions from the packages in which
they're currently housed will do a couple of things:

    * It will create a clear set of dependencies between the labeltools
    functions and all other system packages. (That is, labeltools
    will be a high-level 'consumer' of core packages. We see this
    when we think about the fact that the labeltools functions will
    need to import a number of core packages but that there really
    will be no need now or in future for any existing packages
    to import labeltools.)

    * It will mean that all import statements in labeltools functions
    will be at top level rather than late. This is generally a good
    thing and follows as a corollary from the point above.

    * It will mean that similar code is housed together. That is, that
    all label-like functionality will be housed in a single package
    rather than spread across different package.

    * Interestingly, it will revert to the structure of the ancient 'label' 
    package that was the way in which these functions were first implemented.

As a follow-on to the last point above, it gives me a bit of heartburn
to imagine the package architecture of the system reverting back (in whole
or in part) to the mixed noun- and verb-based naming system used by the
ancient version of the system. For example, would we then revert back
to something like the ancient 'iterate' package and (re)group all 
iteration functions into something like a new iterationtools package?
There are pervasive package architecture onsequences to doing something
like that (even if it would clean up a further set of dependencies).

So it seems like the creation of the labeltools package represents a happy
medium when taking all of these points into consideration: the move
will maintain the noun-only based naming system we now have in place
for all releases of the modern system while also cleaning up intrapackage
dependencies and leaving the out-migrated packages cleaner than they are today.