Removed mathtools.partition_integer_into_thirds().
No function replaces this one. It is simply removed.

Background: the function is too specific for inclusion the API.
Functions that operate on halves of things seem very closely related
to the domain of music composition. Thirds, fourths, fifths and other
divisions are also an important part of what we do. But generalized
functions operating on a generalized understanding of ratios should
be implemented for those cases.

NOTE: this changes end-user syntax and must be noted in the 2.10 release notes.