Regularized symbolic timespan evaluation interface.
There are three symbolic timespan classes:

    timespantools.SymbolicTimespan
    timespantools.SingleSourceSymbolicTimespan
    timespantools.MixedSourceSymbolicTimespan

All three symbolic timespan classes now implement the following
five evaluation methods:

    SymbolicTimespan.get_duration(score_specification)
    SymbolicTimespan.get_score_start_offset(score_specification)
    SymbolicTimespan.get_score_stop_offset(score_specification)
    SymbolicTimespan.get_segment_start_offset(score_specification)
    SymbolicTimespan.get_segment_stop_offset(score_specification)

Enforced on SymbolicTimespan with the abc.abstractmethod decorator.

Note that this brings the symbolic timespan evaluation interface
in line with the selector evaluation interface:

    Selector.get_duration, context_name(score_specification)
    Selector.get_score_start_offset(score_specification, context_name)
    Selector.get_score_stop_offset(score_specification, context_name)
    Selector.get_segment_start_offset(score_specification, context_name)
    Selector.get_segment_stop_offset(score_specification, context_name)

Note that the selector evaluation signature differs from
the symbolic timespan evaluation signature.

It's beyond the scope of this commit message to explain this.
But understanding the difference between the selector and symbolic timespan
evaluation signatures is a good test of understanding the difference
between selectors and symbolic timespans generally.