Changed Request.offset to Request.index globally.
Changed the 'offset' keyword to an 'index' keyword in all Specification request() and set() methods.

Explanation: the term 'offset' should mean only durationtools.Offset everywhere in the codebase.
Prior to this commit, 'offset' was also being used to mean 'j' in expression like [1, 2, 3, 4][j].
Better to refer to 'j' in this situation as an 'index'.

This commit restores a single meaning to 'offset'.