Changed SymbolicTimepoint.addendum to SymbolicTimepoint.offset.
The new name accurately describes what the property does.

Explanation: the previous name of 'addendum' was called into service
because 'offset' was previously used to mean the variable name 
used to index a source. That variable has since changed to 'index'
so that 'offset' may now be used to refer to a temporal offset.