Globally replace 'resolved_value' with 'payload'.
Explanation: the attribute in question refers to the thing carried around
by the various command classes. This is logically a command payload and the
new name helps make that clear. The old name ('resolved_value') was trying
to point to the fact that some payloads are absolute values (like lists of numbers)
while other payloads are score-relative values. But continued work on the interpreter
makes it clear that there are in fact several different times at which
score-relative references will be resolved. So encoding the idea of resolved / unresolved
in a property name is somewhat misleading when reading the code. Better to
think of commands as always having a payload. The *state* of payload depends
on where we are in the process of interpretation.