Update segment specification composer interface.
Details: 

Removed 'truncate' keyword from all setter methods (except division setter methods).
The reason for this is that truncation makes sense only for division setter methods.

Added 'rotation' keyword to all setter methods. The reason for this is that
all setting sources can be treated (implicitly or explicitly) as sequences.
So the 'index', 'count', 'reverse', 'rotation' and 'callback' keywords
make sense as part of the interface to all setter methods.