Integrated AbsoluteRequest objects.
This unifies an important part of the setting interface.

Specifically, it can now be safely observed that all settings carry around a *request* object.

Under the previous regime it was possible for settings to carry around lists, tuples,
numbers and other built-ins. All such values are now wrapped in an absolute request.

In addition to unifying part of the setting interface,
this also has the benefit of providing a type of object-oriented closure
around the values that settings carry around. What that means is that
now every single value carried around by a setting has an outer shell
of the five transform keywords (index, count, reverse, rotation, callback)
that are carried around through the system until just the right moment
during evaluation.

A very fancy type of delayed evaluation.