Restructuring rhythm interpretation to handle rhythm commands one by one.
The current implementation of rhythm interpreation creates all rhythms for
a single voice at one time.

This allows the tiling together of different rhythmic regions within a voice.

But this does not yet allow for rhythm request handling.

Rhythm requests will allow effective cut-and-paste of exact rhythms
from one part of the score to the other.