Incrementally cleaned up rhythm interpretation.
One of the intermediate rhythm region division lists produced during
rhythm interpretation is now modeled as an explicit
RhythmRegionDivisionList object rather than as a built-in
list of divisions.

The general principle is that explicit DivisionList objects
should be used in place of built-in lists of Divsion objects
everywhere it is possible to do so.

The reason for this is that DivsionList object implement
extra list-inspection code that is useful everywhere during
interpretation. So using DivisionList objects in place of
built-in lists limits code duplication between the different
interpreter methods.