Added RhythmExpression to interpretertools package.
RhythmExpression objects model one contiguous chunk of counttime components:
tuplets, ntoes, rests and chords.

The interpretive process of building up the rhythm for a complete
voice of music involves the generation of many different rhythm expressions.
The rhythmic interpretation of a voice completes when enough    
contiguous rhythm expressions exist to account for the entire
duration of the voice.