Request-bearing rhythm commands are now interpreter one after the other.
This is important because rhythm commands will generally have to be
interpreted one after the other to enable back-inspection.

Note that, in fact, *all* commands (at any given 'phase' of interpretation)
have to be interpreted one after the other. Time signature commands are
interpreted one after the other. Division commands are also interpreted
one after the other. So it makes sense for rhythm commands to be
interpreted one after the other, too.

In all cases what's driving this decision is the need for back-inspection.
Meaning the need for later commands to be able to reference earlier commands.