Changed RhythmExpression to OffsetPositionedRhythmExpression.
The class currently inherits from Container and overrides
the way that start_offset and stop_offset are calculated.
Turns out that this is a terrible idea and that start_offset
and stop_offset need to be left alone.

So in a future commit the class will be implemented to *have*
a container (rhythm expression) rather than to *be*
a container.

What that will allow is for the container to preserve
all start_offset and stop_offset calculcation familiar
from the core system while affording user-specified
start- and stop-offset parameters affixed to the wrapper class.