Added private Spanner._reverse_components() method to all spanners.
The method reverses the order of spanner components.

The method is mildly not composer-safe.

The reason for this is that reversing spanner components without
reversing the score order of the components themselves will most
likely result in spanner endpoints formatting backwards.

So this method should called only as part of a complete reversal
routine that also reverses the score-order of components at
the same time.