Added abstract OffsetPositionedExpression class to interpretertools.
OffsetPositionedRhythmExpression now subclasses OffsetPositionedExpression.

The base class opens the way for the implementation of an
OffsetPositionedDivisionExpression class.

Note that the fundamental idea here is that an offset-positioned expression
is a 'floating' piece of (durated) musical material that is to be positioned
at some exact location in the score. So an offset-positioned rhythm expression
might be a run of seven sixteenth notes that is to be positioned to start
at exactly offset 137/8 from the beginning of the score. Likewise, an
offset-positioned division expression might be a division list of
twelve different divisions that is to be positioned to start at exactly
offset 141/16 from the beginning of the score.

So offset-positioned expressions are floating pieces of musical material
that can and should be aligned one after the other to 'tile' the time
of a musical voice.