Migrating from segment-based offsets to score-based offsets.
Interpretation produces the following chain of byproducts:

    * settings
    * commands
    * expressions

Right now expressions work with score-based offsets.
This is proving to be a good thing.

On the other hand, settings and commands work with
segment-based offsets.

The coming series of commits will migrate settings
and commands to work in terms of score-based offsets
instead of segment-based offsets.

This will unify settings, commands and expressions.

This will also time operations everywhere during interpretation.

This commit supplies the Command class with a score_based_duration property
to be used in comparison with the (implicitly segment-based) 'duration' property.
The goal is for these two properties always to return the same value.
Once that is the case, the interim score_based_duration property will be removed.