Added optional 'score_specification' and 'context_name' keywords to TimespanInequality.__call__().
Background: many methods bound to specificationtools-related classes accept
'score_specification' and 'context_name' input parameters. (Sometimes the
parameters are mandatory and sometimes they are optional.) Methods that
accept this pair of input parameters do so as a way of allowing a type
of evaluation context to be passed into the method. (This is similar to
passing locals() or globals() into other functions.)

The current commit extends timespan inequality evaluation optionally to
accept this pair of parameters. Passing these values into a timespan
inequality at call-time enables the inequality to evaluate itself with
respect to the score specification and context name provided.