Reimplemented DivisionSelector.get_score_start_offset() and DivisionSelector.get_score_stop_offset().
The methods are now implemented as wrappers around DivisionSelector.get_score_offsets().

DivisionSelector.get_score_offsets() is now implemented in terms of
DivisionSelector.get_selected_objects().

This means that the offset resolution process is now this:

    * find all divisions for the voice in question
    * filter according to the division selector timespan inequality
    * apply the start_identifer and stop_identifier slice indices

So the process comprises three steps and is reductive at each step.

This commit marks the first time that the division selector timespan inequality
has been consulted. This is a powerfully general move. This means that *all*
timespan selectors now work for selecting divisions.

The pattern will be generalized to *all* selectors. At that point all timespan
inequalities will be usable with all selectors.