Established canonical request type ordering.
The order is:

    * AbsoluteRequest
    * CommandRequest
    * MaterialRequest

These are the three types of request currently implemented.

Several different methods branch on request type. The order
of these if-else branches is arbitrary. This commit
regularizes the order of request type tests to that
given above.

Notionally the ordering proceeds from least number of interpretive
steps (because absolute requests resolve directly) to greatest
number of interpretive steps (because command requests require
time-positioned lookup while material requests require time-positioned
lookup with the addition of the retrieval of fully instantiated
musical material.)