Reimplemented division material request resolution.
As of this commit, division material requests return 
a list of zero or more offset-positioned division expressions.

Prior to this commit, division material requests returned
a single offset-positioned division expression.

The new implementation is considerably better than
the older implementation. The reason for this is that the 
new implementation *preserves breaks* between division regions.

Doing this work has considerably extended TimespanInventory functionality.
The public methods now available on TimespanInventory are these:

    TimespanInventory.adjust_to_stop_offset(stop_offset)
    TimespanInventory.delete_material_that_intersects_timespan(timespan)
    TimespanInventory.get_timespans_that_satisfy_inequality(timespan_inequality)
    TimespanInventory.keep_material_that_intersects_timespan(timespan)
    TimespanInventory.repeat_to_stop_offset(stop_offset)
    TimespanInventory.reverse(stop_offset)
    TimespanInventory.translate_timespans(stop_offset)
    TimespanInventory.trim_to_start_offset(stop_offset)
    TimespanInventory.trim_to_stop_offset(stop_offset)
    
This is a powerful pattern that should be integrated widely.