Completed exemplum [X3].
Explanation:

The large-scale tests driving specificationtools development were written
by hand at the beginning of June or July and kept on a couple of sheets of loose-leaf paper.

These original tests were written in four series: X, A, B, C. 

The number of tests per series is as follows:

    X:  12 tests
    A:  6 tests
    B:  6 tests
    C:  6 tests

These tests can be labelled [X1], [X2], ..., [C5], [C6].

Each test shows what amounts to a higher-level canonical use of the system
and so therefore constitutes a type of exemplum (or schematic example).
This is what is meant when exempla show up in the commit messages and 
in the test doc strings.

It's taken months of continuous development to implement the first three 
of the exempla and as of this commit only [X1], [X2], [X3] are tested.

The other exempla will add to the system soon. And existing test cases will
be organized to make the exemplum series apparent.